깃 충돌 해결
